requirements.txt
dnspython==1.16.0
Flask==1.1.1
flask_restplus==0.13.0
Werkzeug==0.16.0
